requirements.txt
pygame==2.5.2
Pillow==9.5.0
pyinstaller==6.3.0 